requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
cloudinary==1.41.0
Django==5.1.2
django-cloudinary-storage==0.3.0
django-cors-headers==4.5.0
djangorestframework==3.15.2
gunicorn==23.0.0
idna==3.10
packaging==24.1
pillow==11.0.0
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
tzdata==2024.2
urllib3==2.2.3
whitenoise==6.7.0
